set type information for chainer type check

+++ CharRNN.py
@@ -17,8 +17,8 @@ class CharRNN(FunctionSet):
             param[:] = np.random.uniform(-0.08, 0.08, param.shape)
 
     def forward_one_step(self, x_data, y_data, state, train=True, dropout_ratio=0.5):
-        x = Variable(x_data, volatile=not train)
-        t = Variable(y_data, volatile=not train)
+        x = Variable(x_data.astype(np.int32), volatile=not train)
+        t = Variable(y_data.astype(np.int32), volatile=not train)
 
         h0      = self.embed(x)
         h1_in   = self.l1_x(F.dropout(h0, ratio=dropout_ratio, train=train)) + self.l1_h(state['h1'])
@@ -31,7 +31,7 @@ class CharRNN(FunctionSet):
         return state, F.softmax_cross_entropy(y, t)
 
     def predict(self, x_data, state):
-        x = Variable(x_data, volatile=True)
+        x = Variable(x_data.astype(np.int32), volatile=True)
 
         h0      = self.embed(x)
         h1_in   = self.l1_x(h0) + self.l1_h(state['h1'])
